Add method #respond-to-request
This method should receive a client reader and client writer and have no
return value. The request read in from the client reader should be
parsed. Informed by the request, a response should be formed with the
following information:
 - the host name
 - the host port
 - the request path
That request should be sent to the #println method of the client writer.